pyproject.toml
[project]
name = "congress-press"
version = "0.1.0"
description = "Daily collection of congressional press releases with full text"
requires-python = ">=3.12"
dependencies = [
    "python-statement @ git+https://github.com/dwillis/python-statement.git",
    "newspaper4k>=0.9.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["congress_press"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
